Remove old external AX automation, update docs for injection approach
- Delete MacWhisperController.swift (replaced by InjectedMacWhisperController)
- Delete AccessibilityHelper.swift (only used by old controller)
- Extract AXError enum into its own file (still used by injection controller)
- Remove Accessibility permission requirement (injection is in-process, no TCC needed)
- Update PermissionManager to only check Screen Recording
- Update OnboardingView to show Screen Recording as the only permission
- Update AppDelegate permission logic to only require Screen Recording
- Update README with injection architecture, recording workflow, and corrected permissions
- Update FaceTimeDetector comment to reference new controller
