+++ Sources/UI/OnboardingView.swift
@@ -2,47 +2,35 @@ import SwiftUI
 
 struct OnboardingView: View {
     let permissionManager: PermissionManager
-    @State private var accessibilityGranted = false
     @State private var screenRecordingGranted = false
     @State private var checkTimer: Timer?
-
-    private var allGranted: Bool {
-        accessibilityGranted && screenRecordingGranted
-    }
 
     var body: some View {
         VStack(alignment: .leading, spacing: 16) {
             Text("MacWhisperAuto Setup")
                 .font(.headline)
 
-            Text("The following permissions are required for meeting detection and recording automation.")
+            Text(
+                "Screen Recording permission improves meeting detection accuracy."
+                + " Recording automation works without any permissions (via DYLD injection)."
+            )
                 .font(.caption)
                 .foregroundStyle(.secondary)
 
             PermissionRow(
-                title: "Accessibility",
-                description: "Required to control MacWhisper's recording buttons",
-                isGranted: accessibilityGranted,
-                action: {
-                    permissionManager.promptAccessibility()
-                    permissionManager.openAccessibilitySettings()
-                }
-            )
-
-            PermissionRow(
                 title: "Screen Recording",
-                description: "Required to detect meeting windows by title",
+                description: "Recommended for window-title-based meeting detection",
                 isGranted: screenRecordingGranted,
                 action: {
                     permissionManager.openScreenRecordingSettings()
                 }
             )
 
-            if allGranted {
+            if screenRecordingGranted {
                 HStack {
                     Image(systemName: "checkmark.circle.fill")
                         .foregroundStyle(.green)
-                    Text("All permissions granted. Ready to detect meetings.")
+                    Text("Ready to detect meetings and record automatically.")
                         .font(.caption)
                         .foregroundStyle(.green)
                 }
@@ -80,7 +68,6 @@ struct OnboardingView: View {
 
     private func refreshPermissions() {
         let perms = permissionManager.checkAll()
-        accessibilityGranted = perms[.accessibility] ?? false
         screenRecordingGranted = perms[.screenRecording] ?? false
     }
 }
